get authors v1
Hacked version that runs a duplicate article download. The first in getcontent and the second in getauthor.

v0.4 going to amalgamate the two methods. Also tailor feedparser

Starting with a fresh database. Dropping tables was resulting in a high file size

DB renamed as a secondary (cleaned) one is needed later on.